refactor: move internal functions in compiler to separate namespace
