+++ oberon0/Expressions/ExpressionRepository.cs
@@ -9,8 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.Composition;
 using System.Composition.Hosting;
-using Oberon0.Compiler.Definitions;
-using Oberon0.Compiler.Expressions.Functions;
+using Oberon0.Compiler.Expressions.Functions.Internal;
 using Oberon0.Compiler.Expressions.Operations.Internal;
 using Oberon0.Compiler.Types;
