+++ oberon0/Expressions/Functions/FunctionStringLength.cs
@@ -8,10 +8,10 @@
 using JetBrains.Annotations;
 using Oberon0.Compiler.Definitions;
 using Oberon0.Compiler.Expressions.Constant;
-using Oberon0.Compiler.Expressions.Functions;
+using Oberon0.Compiler.Expressions.Functions.Internal;
 using Oberon0.Compiler.Expressions.Operations.Internal;
 
-namespace Oberon0.Compiler.Expressions.Operations;
+namespace Oberon0.Compiler.Expressions.Functions;
 
 /// <summary>
 ///     Handle "Length(STRING)".
@@ -20,7 +20,8 @@ namespace Oberon0.Compiler.Expressions.Operations;
 /// <remarks>This function is some kind of exception as - usually takes one parameter. The second is handled as a dummy</remarks>
 [InternalFunction("INTEGER Length(STRING)")]
 [UsedImplicitly]
-internal class FunctionStringLength: IInternalFunction {
+internal class FunctionStringLength : IInternalFunction
+{
 
     public Expression Operate(FunctionCallExpression e, Block block, InternalFunctionMetadata functionMetadata)
     {
