+++ oberon0/Expressions/Functions/Internal/IInternalFunction.cs
@@ -5,9 +5,17 @@
 // --------------------------------------------------------------------------------------------------------------------
 #endregion
 
+
+#region copyright
+// --------------------------------------------------------------------------------------------------------------------
+// Copyright (c) Stephen Reindl. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+// --------------------------------------------------------------------------------------------------------------------
+#endregion
+
 using Oberon0.Compiler.Definitions;
 
-namespace Oberon0.Compiler.Expressions.Functions;
+namespace Oberon0.Compiler.Expressions.Functions.Internal;
 
 /// <summary>
 ///     Implementations of this interface are responsible to convert data from type a to b.
